Add the column and row prop to span
